Introduce RolloutConfig
